pyproject.toml
[project]
name = "sales-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.83.0",
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
